Further edits on the HOWTO     using inetd doesn't seem to work, so use tcpserver
Added test_comm and hostcheck.pl to MANIFEST and specfile so that
an installation from this src rpm can be tested with a comm and
an injector.

Changed around everything so filehandles are used instead of
fileglobs. Not everywhere yet, but places where it makes a
difference.

mcfeely-spawn and mcfeely-manage just didn't want to talk to one
another when mcfeely-spawn was talking on STDOUT and STDIN so I
changed it to use the same sort of pick up it's file descriptors
from the command line that mcfeely-manage uses.

There is still a problem:

 mcfeely-manage is leaving open the /opt/mcfeely/queue/newj
 directory and eventually the system runs out of files and gets
 all hosed up.

 for some reason when a task passes through from mcfeely-manage
 to mcfeely-spawn, the newj directory is scanned _many_ times.
 The entire scan_jobs routine is called many times and the
 do_select loop is passed through as well.

 so things work for about 10 tasks injections and then it gets
 upset