further explications of how to make it run in the HOWTO. needs work. not friendly, relies on tcpserver, doesn't say anything about initscripts or those sorts of things
adjustments made in some of the plog messages to include the name of
the comm in the output, this helps to make the log more readable,
especially in the case of the oh so helpful, 'program not found' or
whatever it is

do_select now reads a byte from the trigger to prever the select loop
of the trigger from going over and over and over again (for a finite
but noticable amount of time)

jobs.pl uses IO::Dir for scanning the jobs directory, requiring
changes in files.pl as well.

mcfeely-spawn now incorporates the Spawner but it does it in a way
that is not paricularly elegant: basically it incorporates the Spawner
straight up.

this is working with simple hostcheck.pl single task injections of
test_comm

hostcheck.pl and test_comm now get installed by the rpm, this means
mcfeely-client and mcfeely-comms are going to conflict, so when I
merge this I will need to remove test_comm from mcfeely-comms

still needed:

- some comments explaining the pipes
- some comments explaining the loops
- in some places the code logic could be expanded to make it a bit
  more clear
- I'm not sure, but I think there are a few places where OO
  filehandling is not being used

N.B: this stuff requires late model IO::* support to get access to the
blocking() method.