Adjustments to all the c code that contains a main() so that they are 'int main()' instead of 'void main()'. Shuts up some compiler warnings. Created some more that I was able to shut up by putting some not needed but clean exit()s at the end of the main block.
Pointed people to the HOWTO doc from the README doc.

However, the HOWTO and the README are not included in the RPM so
I noted in the TODO that we need to make some effort to get some
of the docs into the distribution.

created and added the serverslam.pl test script which slams
mcfeely servers, just like it says. not (yet) well documented.

still to do:

- comments in the stuff that i've changed
- HOWTO doc to be pushed up to web during 'make update'
- find the bug that serverslam.pl reveals with stranded tasks
- merge, eventually