some error trapping in mcfeely-spawn and changes to how those errors are reports so in the case where we can't fork the other end knows about through a sensical return down the pipe, not something that won't work at all
moved all blocking() calls back to mcfeely-start, except for the
trigger in do_select

radically altered read_results so that it can read all the data
that is waiting on the pipe, not just the next one. _huge_
performance increase when there are tons of tasks, however it
can't work as it is currently done because it us 0x4 as $/ and
that value can show up in the binary packed return from
mcfeely-spawn. I wonder if it isn't worthwhile just to get rid of
the packing and separate on \0 or something like that. Adrian
went through this before, manifests itself as the bad data being
returned on the pipe:

starting
13368 info: slamserver job 2
13606 info: slamserver job 10
14895 info: slamserver job 10
15117 info: slamserver job 17
15117:15108 (test_comm) starting transfer to saliva.kiva.net
14895:14852 (test_comm) starting transfer to saliva.kiva.net
13606:13572 (test_comm) starting transfer to saliva.kiva.net
13368:13316 (test_comm) starting transfer to saliva.kiva.net
heading for the while loop for reading srr ####################
got 59, 97, rg 6 is present
heading for the while loop for reading srr ####################
got 58, 97, rg 12 is present
heading for the while loop for reading srr ####################
got 53, 97, rg 21 is present
heading for the while loop for reading srr ####################
got 52, 97, rg 1 is present

I'm committing so I can work on this some more from home later.