Oi!
Changed around the pipe reading in read_results so that we get
all the data from a read instead of doing <> operations. Made it
so there is a buffer that stores incomplete data. It doesn't get
used except in unusual circumstances but can probably be used
later when we want to be a little better about using less
resources.

Made some adjustments in mcfeely-spawn so that data is gathered
and stored on variables that are correctly scoped. In previous
versions the variables could get _huge_ and strange things would
happen. I think.

Change the <> that gets data from mcfeely-ttpc in mcfeely-spawn
to be a read, allows us to have multi-line reports. or at least
it should.

put an END() in mcfeely-spawn to try and catch situations where
it just ups and dies.

put a reaper() in mcfeely-manage to catch when mcfeely-spawn
dies.

Added a few comments here in there. That's what I planned to do
most of today but kept finding more things worth fixing.

Things discussed today but not done:

-  making the read after select in mcfeely-spawn more clear
-  constraining the read in read_results
-  constraining how many tasks get launched

It is not immediately clear to me what the state of current bugs
are. On a clean queue I seem to do okay (most of the time?).
On a messed up queue, things get weird, as one might expect, but
that isn't good: the system should be resilient enough to
actively fail with knowledge, instead of use bad data.