mcfeely-manage now takes as it's first arg the pid of mcfeely-spawn, the child so that the repeaer() can identify which child has just been waited on
serverslam changes ARGV[4] 'mail' to args 4 and 5 fnot and snot
respectively so that that you can pass just one of those if you
want.

This latest commit is close but no cigar. I think it needs
Adrian's eyes on it, given his current thinking about resource
startvation. It is only under extreme circumstances that things
go wrong:

-  sometimes mcfeely-spawn will core dump and die
-  sometimes a task will be half completed in that the info file
   will have already been unlinked, and the job is considered
   done but something will try to unlink it anyway
-  tasks will get stranded, this seems to be an indication that
   in the transferring of data a tasks gets dropped

The new code is easy to install and test. Check it out, make rpm,
install the rpms, create the secret files, tcpserver mcfeely-ttpd
in one window and bin/mcfeely-start in another and then beat on
it with serverslam.pl