Multiple cleanup type changes.
First, mcfeely-queue:

made changes in copy_bytes.c to ensure that if for some reason
the sending pipe blocks, we still read information correctly,
this allows us to read large numbers of tasks.

that change revealed a problem in Job.pm where the pipes to
mcfeely-queue need to be set autoflush or things go wrong.

that change revealed a problem (not a bug as such) in
mcfeely-queue.h where the ino[] array was limited to 256 entries.
I upped this to 16384, which may seem excessive, but is roomy
enough for injecting 10000 tasks in one job, which was my test
limit.

Moved on to try and make mcfeely-manage and it's associated files
a bit easier to debug. Created McFeely::Log and McFeely::Const so
that plog(), bail() and the TASK_NDEPS and associates are
available more easily. The *.pl files have been changed to 'use
McFeely::Log' and 'use McFeely::Const' so a perl -cw tasks.pl can
be successful.

do_select has been merged into mcfeely-manage

serverslam.pl produces some output that says how many tasks were
injected per each job.

removed const.pl and log.pl from Makefile, MANIFEST and specfile.
Added Log.pm and Const.pm

did a

  for i in *.pl ; do perl -Mstrict -cw $i ; done

and fixed what scoping problems I could find. There is a fair bit
of confusion in jobs.pl scan_jobs where $task is used both as a
list and a hash and I can't for the life of me figure out the
scoping stuff there

Most of the *.pl files won't go strict, but that is to be
expected given the way they are used. I considered rolling them
all into mcfeely-manage (so strictness could be had) but that
would make for a very large single script. Suggestions are
welcome.

I'm on a branch called queue.

I'm going to work on the ttpd on inetd issue now.