Preliminary run at getting install and uninstall triggers to work seems to be going well.
Got docs to be included in mcfeely-common rpm. Not including all docs
because some are not complete enough for shipping...but adding further
docs in the future should be trivial.

This is not a complete fix of the tasks I'm taking with McFeely, but
it's a big step forward so it's time to commit it.