mcfeely-logger: the syslogging program that syslogs it's input to   local6
mcfeely.logrotate: for use by the rpm, drops a mcfeely file in
  /etc/logrotate.d

mcfeely-init changed to reflect the new logger. also now checks
for the presence of the /var/lock/subsys/mcfeely-start file and
will not start if it is there

mcfeely-manage and mcfeely-spawn have been adjusted to provide
better exiting. mcfeely-spawn ignores TERM. mcfeely-manage, on
receipt of TERM, stops sending tasks and waits for responses on
active tasks from mcfeely-spawn

other changes are to reflect file additions.